Fix $GTK_MODULES race condition
Partially revert r337: Keep the upstart job as it is to let it take care of
setting the env var in upstart, and only set the systemd+dbus variable in the
systemd unit. We cannot order systemd units against upstart job, thus cannot
enforce that upstart jobs run after unity-gtk-module.service, which caused the
variable to be wrong in a lot of upstart-launched programs.