meson: Move gdk-pixbuf to Requires.private in appmenu-glib-translator.pc
`Requires` makes it so the dependency must be fully available in the build environment of projects depending on `appmenu-glib-translator`. In ecosystems with package isolation like Nix, this requires adding the transitive dependency to the environment explicitly, or propagating the dependency.

However, gdk-pixbuf is not currently used by the translator. And even if the use of the dependency was not commented out, it would still not be exposed in public headers so putting it to `Requires.private` is sufficient.